refactor: split run's sandbox pipeline into focused modules under lib/sandbox/
