+++ run/src/lib/sandbox/oci-config.ts
@@ -1,17 +1,7 @@
-import { execFileSync } from "node:child_process";
-import {
-  writeFileSync,
-  mkdtempSync,
-  rmSync,
-  readFileSync,
-  mkdirSync,
-  chmodSync,
-  existsSync,
-} from "node:fs";
-import { join, dirname } from "node:path";
-import { fileURLToPath } from "node:url";
-import { buildDockerCpArgs } from "../../../core/lib/docker/cp-args.ts";
-import { errorMessage } from "../../../core/lib/general/error-message.ts";
+import { writeFileSync, existsSync } from "node:fs";
+import { join } from "node:path";
+import type { HasMounts, OciSpec, BuiltOciSpec, HostMount } from "./types.ts";
+import { SANDBOX_SCRATCH_BASE } from "./scratch-dir.ts";
 // Sensitive /proc paths masked with /dev/null. runc's own `runc spec`
 // default already masks /proc/kcore, /proc/keys, and /proc/timer_list
 // (among others) and leaves /proc/sysrq-trigger merely read-only —
@@ -23,64 +13,7 @@ import { errorMessage } from "../../../core/lib/general/error-message.ts";
 // build-test-bundle.sh — a bash/jq stand-in for this same function, used
 // by the Mac dev loop — has a single source of truth to read the same
 // list from instead of hand-duplicating it.
-import EXTRA_MASKED_PROC_PATHS from "../../scripts/extra-masked-proc-paths.json" with { type: "json" };
-
-// rollup's cjs output doesn't convert import.meta.dirname (it silently
-// becomes undefined), so use this form instead.
-const __dirname = dirname(fileURLToPath(import.meta.url));
-
-// Base directory for each run's scratch dir (OCI bundle + the host-`/`
-// rootfs bind-mount). Deliberately under /var/tmp rather than os.tmpdir():
-// the rootfs bind must live somewhere that is never one of the sandbox's
-// writable exceptions (workdir/home/tmp/RUNNER_TEMP/writablePaths),
-// otherwise the recursive writable rbind of that path would re-expose the
-// whole host `/` as a second, writable copy inside the sandbox. /var/tmp
-// itself is 1777 (writable by the non-root runner user) and execable, so
-// this own subdirectory inherits that without needing root to create it.
-// buildOciConfig fails closed if a step's `writable:` input tries to list
-// this directory (or an ancestor of it) as writable — see
-// assertScratchBaseNotWritable.
-const SANDBOX_SCRATCH_BASE = "/var/tmp/buildcage";
-
-interface MountEntry {
-  destination: string;
-  type?: string;
-  source?: string;
-  options?: string[];
-}
-
-interface OciSpec {
-  mounts: MountEntry[];
-  linux: {
-    maskedPaths?: string[];
-    readonlyPaths?: string[];
-    namespaces: { type: string; path?: string }[];
-    seccomp?: unknown;
-  };
-  root?: unknown;
-  process: Record<string, unknown>;
-}
-
-// buildOciConfig's actual, guaranteed-populated output shape — narrower than
-// the general OciSpec above, which also stands in for runc's raw, more
-// loosely-known `runc spec` input.
-interface BuiltOciSpec extends OciSpec {
-  root: { path: string; readonly: boolean };
-  process: Record<string, unknown> & {
-    args: string[];
-    env: string[];
-    cwd: string;
-    user: { uid: number; gid: number };
-    capabilities: unknown;
-    noNewPrivileges: boolean;
-  };
-  linux: OciSpec["linux"] & { maskedPaths: string[]; readonlyPaths: string[]; seccomp: unknown };
-}
-
-interface HostMount {
-  mountPoint: string;
-  fsType: string;
-}
+import EXTRA_MASKED_PROC_PATHS from "../../../scripts/extra-masked-proc-paths.json" with { type: "json" };
 
 /**
  * Write the user-supplied `run:` input to an executable script file.
@@ -92,105 +25,6 @@ export function writeRunScript(runInput: string, dir: string): string {
   const content = runInput.startsWith("#!") ? runInput : `#!/bin/sh\nset -e\n${runInput}\n`;
   writeFileSync(scriptPath, content, { mode: 0o700 });
   return scriptPath;
-}
-
-/**
- * Generate runc's own default OCI bundle config via `runc spec` (run in
- * `bundleDir`, which is where it writes `config.json`). Used as the
- * starting point for buildOciConfig rather than hand-writing the full
- * spec from scratch, so the baseline mounts/masked-paths/rlimits stay
- * exactly what runc itself considers a sane default for its own version,
- * and buildOciConfig only needs to override/extend the handful of fields
- * this sandbox actually cares about.
- */
-export function generateBaseOciSpec(runcPath: string, bundleDir: string): OciSpec {
-  execFileSync(runcPath, ["spec"], { cwd: bundleDir });
-  return JSON.parse(readFileSync(join(bundleDir, "config.json"), "utf8"));
-}
-
-/**
- * Extract runc and gen-seccomp-profile from the proxy image into this run's
- * own `destDir` (its per-step scratch dir), then resolve the base OCI spec
- * and the seccomp profile from them. Run once per `run:` step; each
- * invocation is independent, and everything written here is torn down with
- * the scratch dir (see withScratchDir / cleanupScratchDir).
- *
- * Both binaries ship inside the proxy image and are pulled onto the host via
- * `docker cp`, then run natively there (not `docker exec`) since the seccomp
- * profile's content depends on the real host kernel/arch -- see
- * gen-seccomp-profile/main.go. gen-seccomp-profile is only needed transiently
- * to resolve the profile, so it's removed once read; runc stays for `runc run`.
- */
-export interface ExtractRuncBootstrapOptions {
-  containerName: string;
-  destDir: string;
-}
-
-export interface RuncBootstrap {
-  runcPath: string;
-  seccompProfile: unknown;
-  baseSpec: OciSpec;
-}
-
-export function extractRuncBootstrap({
-  containerName,
-  destDir,
-}: ExtractRuncBootstrapOptions): RuncBootstrap {
-  const runcPath = join(destDir, "runc");
-  const genSeccompProfilePath = join(destDir, "gen-seccomp-profile");
-  execFileSync(
-    "docker",
-    buildDockerCpArgs({
-      containerName,
-      containerPath: "/opt/buildcage/bin/runc",
-      hostPath: runcPath,
-    }),
-  );
-  execFileSync(
-    "docker",
-    buildDockerCpArgs({
-      containerName,
-      containerPath: "/opt/buildcage/bin/gen-seccomp-profile",
-      hostPath: genSeccompProfilePath,
-    }),
-  );
-  chmodSync(runcPath, 0o755);
-  chmodSync(genSeccompProfilePath, 0o755);
-  const seccompProfile = JSON.parse(execFileSync(genSeccompProfilePath, { encoding: "utf8" }));
-  const baseSpec = generateBaseOciSpec(runcPath, destDir); // writes config.json into destDir (overwritten later by writeOciConfig)
-  rmSync(genSeccompProfilePath); // only needed to resolve seccompProfile above
-
-  return { runcPath, seccompProfile, baseSpec };
-}
-
-/**
- * Pure: extract {mountPoint, fsType} for every line of raw
- * /proc/self/mountinfo content. Format (space-separated fields):
- *   ID PARENT-ID MAJOR:MINOR ROOT MOUNT-POINT OPTIONS [OPT-FIELDS...] - FSTYPE SOURCE SUPER-OPTIONS
- * The mount point is always field 5 (index 4); the filesystem type is
- * always the field right after the literal "-" separator, regardless of
- * how many optional fields precede it.
- */
-export function parseMountinfo(mountinfoContent: string): HostMount[] {
-  return mountinfoContent
-    .split("\n")
-    .filter(Boolean)
-    .map((line) => {
-      const fields = line.split(" ");
-      const dashIndex = fields.indexOf("-");
-      return { mountPoint: fields[4], fsType: fields[dashIndex + 1] };
-    });
-}
-
-/**
- * Reads the real host mount table. Node runs directly on the runner host,
- * not inside any namespace, so this is exactly the mount table
- * run-isolated.sh's `mount --rbind /` will duplicate into rootfsBindDir a
- * moment later (see buildOciConfig's readonlyPaths handling for why this
- * matters).
- */
-export function listHostMounts(): HostMount[] {
-  return parseMountinfo(readFileSync("/proc/self/mountinfo", "utf8"));
 }
 
 /**
@@ -238,10 +72,6 @@ export function computeReadonlyHostMounts(
  * cgroup v1 or v2 hierarchy, so matching by destination path covers both
  * without needing to special-case a literal "cgroup2" fstype name).
  */
-export interface HasMounts {
-  mounts: MountEntry[];
-}
-
 export function freshMountDestinationsFrom(baseSpec: HasMounts): Set<string> {
   return new Set(baseSpec.mounts.map((m) => m.destination));
 }
@@ -483,190 +313,4 @@ export function writeResolvConf(dns: string, dir: string): string {
   const resolvConfPath = join(dir, "resolv.conf");
   writeFileSync(resolvConfPath, `nameserver ${dns}\n`, { mode: 0o644 });
   return resolvConfPath;
-}
-
-/**
- * Run the user's command inside the isolated sandbox via run-isolated.sh
- * (invoked with `sudo -n`, since setting up namespaces/veth/iptables/the
- * rootfs bind-mount requires root). Returns the exit code of the isolated
- * command — never throws for a non-zero exit, since that's the user's
- * command failing, not this function.
- *
- * uid/gid, capabilities, mounts, and env are entirely described by
- * `config.json` (see buildOciConfig) — run-isolated.sh only needs enough
- * to set up networking and the rootfs bind-mount before handing off to
- * `runc run`.
- */
-export interface RunIsolatedOptions {
-  runcPath: string;
-  proxyPid: number;
-  bundleDir: string;
-  containerId: string;
-  netnsName: string;
-  rootfsBindDir: string;
-  gateway: string;
-  dns: string;
-  targetIp: string;
-}
-
-export function runIsolated({
-  runcPath,
-  proxyPid,
-  bundleDir,
-  containerId,
-  netnsName,
-  rootfsBindDir,
-  gateway,
-  dns,
-  targetIp,
-}: RunIsolatedOptions): number {
-  const runIsolatedShPath = join(__dirname, "..", "scripts", "run-isolated.sh");
-
-  const args = [
-    "-n",
-    "--",
-    runIsolatedShPath,
-    "--proxy-pid",
-    String(proxyPid),
-    "--runc",
-    runcPath,
-    "--bundle",
-    bundleDir,
-    "--container-id",
-    containerId,
-    "--netns-name",
-    netnsName,
-    "--rootfs-bind-dir",
-    rootfsBindDir,
-    "--gateway",
-    gateway,
-    "--dns",
-    dns,
-    "--target-ip",
-    targetIp,
-  ];
-
-  try {
-    execFileSync("sudo", args, { stdio: "inherit" });
-    return 0;
-  } catch (e) {
-    // A non-zero exit from the isolated command (or run-isolated.sh itself)
-    // surfaces here as an ExecException; e.status is the actual exit code.
-    // e.status is null if the process was killed by a signal.
-    const status = (e as { status?: number | null }).status;
-    return typeof status === "number" ? status : 1;
-  }
-}
-
-/**
- * Pure: mount points from raw /proc/self/mountinfo content that are
- * nested under `dir` (including `dir` itself), deepest-path-first so a
- * caller can safely unmount children before their parents.
- */
-export function parseMountsUnder(mountinfoContent: string, dir: string): string[] {
-  const prefix = dir.endsWith("/") ? dir : `${dir}/`;
-  return parseMountinfo(mountinfoContent)
-    .map(({ mountPoint }) => mountPoint)
-    .filter((mountPoint) => mountPoint === dir || mountPoint.startsWith(prefix))
-    .sort((a, b) => b.length - a.length);
-}
-
-/**
- * Force-detaches any mount points still nested under `dir` before it's
- * recursively deleted. This is the safety net for rootfsBindDir (a
- * `mount --rbind /` of the entire host filesystem — see main.ts) surviving
- * past run-isolated.sh's own cleanup trap: if that trap never runs (e.g.
- * run-isolated.sh itself is SIGKILL'd, which bypasses traps entirely) or
- * its `umount -R` fails (EBUSY), a plain recursive delete of `dir` would
- * otherwise walk straight through the still-live bind-mount and delete
- * the real files on the host it points at, not a sandboxed copy. `-l`
- * (lazy) detaches each mount from the namespace immediately regardless of
- * busy references, so this step itself can't hang or fail the way a
- * normal (non-lazy) unmount could.
- */
-function unmountAllUnder(dir: string): void {
-  let mountPoints;
-  try {
-    mountPoints = parseMountsUnder(readFileSync("/proc/self/mountinfo", "utf8"), dir);
-  } catch {
-    return;
-  }
-  for (const mountPoint of mountPoints) {
-    try {
-      execFileSync("sudo", ["umount", "-R", "-l", mountPoint], {
-        stdio: ["ignore", "ignore", "pipe"],
-      });
-    } catch (e) {
-      console.log(`::warning::Failed to unmount ${mountPoint} before cleanup: ${errorMessage(e)}`);
-    }
-  }
-}
-
-/**
- * Removes the scratch dir, retrying on EBUSY. A lazy unmount (see
- * unmountAllUnder) detaches a mount from the path-resolution tree
- * immediately -- it stops appearing in /proc/self/mountinfo right away --
- * but the kernel's underlying teardown of that now-orphaned mount can
- * still lag behind by a short, bounded window, which can make a
- * directory rmSync is about to delete spuriously report EBUSY even
- * though it's no longer listed as a mountpoint at all. Resolves on the
- * very next attempt after a brief wait.
- */
-function removeScratchDir(dir: string): void {
-  const maxAttempts = 5;
-  for (let attempt = 1; attempt <= maxAttempts; attempt++) {
-    try {
-      rmSync(dir, { recursive: true, force: true });
-      return;
-    } catch (e) {
-      if ((e as NodeJS.ErrnoException).code !== "EBUSY" || attempt === maxAttempts) throw e;
-      Atomics.wait(new Int32Array(new SharedArrayBuffer(4)), 0, 0, 200);
-    }
-  }
-}
-
-/**
- * Force-detach anything still mounted under `dir` (the rootfs bind-mount
- * safety net — see unmountAllUnder) and then recursively remove it. Exported
- * so post.ts can reclaim a scratch dir orphaned by a hard kill that bypassed
- * withScratchDir's own finally. No-ops safely when `dir` doesn't exist.
- */
-export function cleanupScratchDir(dir: string): void {
-  unmountAllUnder(dir);
-  removeScratchDir(dir);
-}
-
-/**
- * Absolute path of the scratch dir for a given proxy container, derived
- * deterministically from `containerName` (the `buildcage-proxy-` prefix
- * swapped for `sandbox-`, under SANDBOX_SCRATCH_BASE). Lets the post step
- * reconstruct and reclaim the exact same directory from `STATE_container_name`
- * alone.
- */
-export function scratchDirFor(containerName: string): string {
-  return join(SANDBOX_SCRATCH_BASE, containerName.replace(/^buildcage-proxy-/, "sandbox-"));
-}
-
-/**
- * Create/remove a scratch directory for this step's OCI bundle + run-script.
- * With `containerName` the dir is named deterministically (scratchDirFor) so
- * post.ts can reclaim it after a hard kill; without it a random mkdtemp name
- * is used (unit tests). Cleaned up on every exit path that unwinds — a
- * SIGKILL bypasses this finally, which is exactly what post.ts covers.
- */
-export function withScratchDir<T>(fn: (dir: string) => T, containerName?: string): T {
-  let dir: string;
-  mkdirSync(SANDBOX_SCRATCH_BASE, { recursive: true, mode: 0o755 });
-  if (containerName) {
-    dir = scratchDirFor(containerName);
-    cleanupScratchDir(dir); // clear any stale remnant at this deterministic path (unmount-safe)
-    mkdirSync(dir, { recursive: true, mode: 0o700 });
-  } else {
-    dir = mkdtempSync(join(SANDBOX_SCRATCH_BASE, "sandbox-"));
-  }
-  try {
-    return fn(dir);
-  } finally {
-    cleanupScratchDir(dir);
-  }
 }
