+++ run/src/post.ts
@@ -5,7 +5,7 @@ import { fileURLToPath } from "node:url";
 import * as core from "@actions/core";
 
 import { buildComposeDownArgs } from "../../core/lib/docker/compose-args.ts";
-import { cleanupScratchDir, scratchDirFor } from "./lib/isolated-exec.ts";
+import { cleanupScratchDir, scratchDirFor } from "./lib/sandbox/scratch-dir.ts";
 import { errorMessage } from "../../core/lib/general/error-message.ts";
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
